Code written for LinkIt
The device detects upto a range of 22-24m comfortably even with objects in the way. Sitting in a bus doesn't have much effect on the range.
Range is negligibly more than Edison, like a meter max. So, both have like the same range.

Since scan time is 3 seconds, scans nearby devices comfortably. An issue is that the number of devices being detected after one scan is not very consistent.Taking the MAC addresses after 2-3 scans gives a more accurate number of BT devices in the area.
Also, on the website it said that a scan can detect at max. 10 devices. 
(An Idea, Not Implemented, Subject to discussion) So, thought of creating sets to store the data provided in two three scans and then sending this set to the sd card. 

Another issue is that, let's say there is a person walking on the pavement and using bluetooth, like on his phone or bluetooth headest. The device cannot distinguish between a vehicle and that person.

A problem in the code is that even though data is to be sent to the mbt.txt file on sd card,
an error props up, saying that it can't even open such a file.
On the serial monitor, data is being shown normally. 
Opening the .txt file is not being implementd. Data is not being sent to sd card. Couldn't find a way to resolve the problem.